requirements.txt
PyYAML==6.0.2
requests==2.32.3
pytest==8.3.4
google-analytics-data==0.18.18
